requirements.txt
Flask==3.0.3
tensorflow==2.19.0
numpy==1.26.4
Pillow==10.4.0